fixing issue with files that have multiple methods
Previous versions of XWB had hard-coded the method index to a value of '1', which only works if the MS method is first in the file.  However, if there was another method in the '1' position (e.g. the LC method), then may things would work properly.  This version adds a function GetMSInstrumentMethod(), which looks at the method names and tries to be smart about which one it returns.  NOTE: This includes a newer version of the LuaRawFile.dll, which enables this functionality.